app/requirements.txt
pydantic==2.7
uvicorn==0.30.1
python-dotenv==1.0.1
fastapi==0.111.0
opentelemetry-exporter-jaeger-proto-grpc==1.21.0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-exporter-jaeger-thrift==1.21.0
opentelemetry-exporter-jaeger==1.21.0
 